feat(security): add SecurityConfig with JWT filter chain and public routes
